refactor: move AiChatEngine to promptkt-core, wire REPL to use multimodal models as text chat
AiChatEngine sealed class and asTextChat() adapter moved from tri.promptfx to tri.ai.core
so it can be shared by both promptfx and promptrt-cli.

Added AiModelProvider.chatEngine() and allChatEngines() helpers that resolve both TextChat
and MultimodalChat models (e.g. gpt-4o, Claude 3, Gemini) for use as text chat engines.
The REPL now sees all models that were previously invisible because they are only registered
as MultimodalChat.

Also switched promptrt-cli fat jar from maven-assembly-plugin to maven-shade-plugin with
ServicesResourceTransformer so META-INF/services files from all provider JARs are merged
rather than overwritten, fixing provider discovery at runtime.
