+++ promptfx/promptfx/src/main/kotlin/tri/promptfx/docs/DocumentInsightView.kt
@@ -34,7 +34,7 @@ import tri.ai.prompt.template
 import tri.ai.prompt.trace.AiOutput
 import tri.ai.prompt.trace.batch.AiPromptBatchCyclic
 import tri.promptfx.AiPlanTaskView
-import tri.promptfx.execute
+import tri.ai.core.execute
 import tri.promptfx.PromptFxConfig
 import tri.promptfx.PromptFxGlobals.promptsWithPrefix
 import tri.promptfx.PromptFxModels
