+++ promptfx/promptfx/src/main/kotlin/tri/promptfx/fun/WeatherAiTaskPlanner.kt
@@ -29,10 +29,10 @@ import tri.ai.pips.AiTaskBuilder
 import tri.util.json.jsonMapper
 import tri.ai.prompt.trace.AiOutput
 import tri.ai.prompt.trace.AiOutputInfo
-import tri.promptfx.AiChatEngine
+import tri.ai.core.AiChatEngine
+import tri.ai.core.execute
 import tri.promptfx.ModelParameters
 import tri.promptfx.PromptFxGlobals.lookupPrompt
-import tri.promptfx.execute
 import tri.util.fine
 import tri.util.info
