+++ promptfx/promptfx/src/main/kotlin/tri/promptfx/prompts/PromptScriptView.kt
@@ -32,7 +32,7 @@ import tri.ai.prompt.template
 import tri.ai.prompt.trace.*
 import tri.ai.prompt.trace.batch.AiPromptBatchCyclic
 import tri.promptfx.AiPlanTaskView
-import tri.promptfx.execute
+import tri.ai.core.execute
 import tri.promptfx.PromptFxGlobals.promptsWithPrefix
 import tri.promptfx.PromptFxModels
 import tri.promptfx.TextLibraryReceiver
